moves NonDraggableElement into the DragAndDropComponents folder
