+++ app/components/DragAndDropComponents/NonDraggableElement.js
